allsocs: tifs_sc: Fix ISC structure
This commit fixes ISC structure where
we store the golden config and the
retrieved configuration for CBASS.
RCR or region control register is applicable
to a set of regions together and needs to
common for a register list. The previous
code had a mix of both notions (RCR per region
in struct but being checked only once for a
set of regions) causing array overflow.
This affects all devices. The commit fixes it
for J7x devices.

Fixes: SAFETY_CHECKERS-553

Signed-off-by: Tanu Hari Dixit <thdixit@ti.com>